% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_params_doc.R
\name{default_params_doc}
\alias{default_params_doc}
\title{This function does nothing. It is intended to inherit is parameters'
documentation.}
\usage{
default_params_doc(
  fasta_filename,
  locatome,
  protein_sequence,
  protein_sequences,
  verbose
)
}
\arguments{
\item{fasta_filename}{path to a FASTA file}

\item{locatome}{text similar to a FASTA file, with protein names
and the locations of its amino acids}

\item{protein_sequence}{a protein sequence}

\item{protein_sequences}{one or more protein sequences}

\item{verbose}{set to TRUE for more output}
}
\description{
This function does nothing. It is intended to inherit is parameters'
documentation.
}
\note{
This is an internal function, so it should be marked with
\code{@noRd}. This is not done, as this will disallow all
functions to find the documentation parameters
}
\author{
Richel J.C. Bilderbeek
}
